refactored structure of styled components
